consumindo API Resti com biblioteca Restify

+++ controllers/pagamentos.js
@@ -44,11 +44,11 @@ module.exports = function(app) {
     });
 
     app.post("/pagamentos/pagamento", function(req, res) {
-        var pagamento = req.body;
+        var pagamento = req.body["pagamento"];
         console.log('processando pagamento...');
 
-        req.assert("forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
-        req.assert("valor", "Valor é obrigatório e deve ser um decimal").notEmpty().isFloat();
+        req.assert("pagamento.forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
+        req.assert("pagamento.valor", "Valor é obrigatório e deve ser um decimal").notEmpty().isFloat();
 
         var erros = req.validationErrors();
 
@@ -72,25 +72,65 @@ module.exports = function(app) {
 
                 pagamento.id = result.insertId
 
-                res.location('/pagamentos/pagamento/' + pagamento.id);
+                if (pagamento.forma_de_pagamento == 'cartao'){
+                    var cartao = req.body["cartao"];
+                    console.log(cartao);
 
-                var response = {
-                    dados_do_pagamento: pagamento,
-                    links: [
-                        {
-                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                            rel:"confirmar",
-                            method:"PUT"
-                        },
-                        {
-                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                            rel:"cancelar",
-                            method:"DELETE"
+                    var clienteCartoes = new app.servicos.CartoesClient();
+
+                    clienteCartoes.autoriza(cartao, function(exception, request, response, retorno){
+                        if(exception){
+                            console.log(exception);
+                            res.status(400).send(exception);
+                            return;
                         }
-                    ]
-                }
+                        console.log(retorno);
 
-                res.status(201).json(response);
+                        res.location('/pagamentos/pagamento/' + pagamento.id);
+
+                        var response = {
+                            dados_do_pagamanto: pagamento,
+                            cartao: retorno,
+                            links: [
+                                {
+                                href:"http://localhost:3000/pagamentos/pagamento/"
+                                        + pagamento.id,
+                                rel:"confirmar",
+                                method:"PUT"
+                                },
+                                {
+                                href:"http://localhost:3000/pagamentos/pagamento/"
+                                        + pagamento.id,
+                                rel:"cancelar",
+                                method:"DELETE"
+                                }
+                            ]
+                        }
+                        res.status(201).json(response);
+                        return;
+                    });
+            
+                } else {
+
+                    res.location('/pagamentos/pagamento/' + pagamento.id);
+
+                    var response = {
+                        dados_do_pagamento: pagamento,
+                        links: [
+                            {
+                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                rel:"confirmar",
+                                method:"PUT"
+                            },
+                            {
+                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                rel:"cancelar",
+                                method:"DELETE"
+                            }
+                        ]
+                    }
+                    res.status(201).json(response);
+                }
             }
         });
     });
